game-of-tracing/war_map/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0 
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp